Moved extra code to gulp helpers
